requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
Werkzeug
requests
uuid
markdown2
chromadb
sentence-transformers
numpy
pymupdf
pandas
docx
PyMuPDF
PyPDF2
google-generativeai
openai
npttf2utf
